WIP working version of ground-truth matching, missing documentation
