+++ ultrack/core/linking/processing.py
@@ -146,25 +146,52 @@ def _process(
         next_nodes = [row[0] for row in query]
         next_shift = np.asarray([row[1:] for row in query])
 
-    current_pos = np.asarray([n.centroid for n in current_nodes])
-    next_pos = np.asarray([n.centroid for n in next_nodes], dtype=np.float32)
+    compute_spatial_neighbors(
+        time,
+        config,
+        current_nodes,
+        next_nodes,
+        next_shift,
+        scale=scale,
+        table_name=LinkDB.__tablename__,
+        db_path=db_path,
+        images=images,
+        write_lock=write_lock,
+    )
 
-    n_dim = next_pos.shape[1]
-    next_shift = next_shift[:, -n_dim:]  # matching positions dimensions
-    next_pos += next_shift
+
+def compute_spatial_neighbors(
+    time: int,
+    config: LinkingConfig,
+    source_nodes: List[Node],
+    target_nodes: List[Node],
+    target_shift: ArrayLike,
+    scale: Optional[Sequence[float]],
+    table_name: str,
+    db_path: str,
+    images: Sequence[ArrayLike],
+    write_lock: Optional[fasteners.InterProcessLock] = None,
+) -> pd.DataFrame:
+
+    source_pos = np.asarray([n.centroid for n in source_nodes])
+    target_pos = np.asarray([n.centroid for n in target_nodes], dtype=np.float32)
+
+    n_dim = target_pos.shape[1]
+    target_shift = target_shift[:, -n_dim:]  # matching positions dimensions
+    target_pos += target_shift
 
     if scale is not None:
         min_n_dim = min(n_dim, len(scale))
         scale = scale[-min_n_dim:]
-        current_pos = current_pos[..., -min_n_dim:] * scale
-        next_pos = next_pos[..., -min_n_dim:] * scale
+        source_pos = source_pos[..., -min_n_dim:] * scale
+        target_pos = target_pos[..., -min_n_dim:] * scale
 
     # finds neighbors nodes within the radius
     # and connect the pairs with highest edge weight
-    current_kdtree = KDTree(current_pos)
+    current_kdtree = KDTree(source_pos)
 
     distances, neighbors = current_kdtree.query(
-        next_pos,
+        target_pos,
         # twice as expected because we select the nearest with highest edge weight
         k=2 * config.max_neighbors,
         distance_upper_bound=config.max_distance,
@@ -173,8 +200,8 @@ def _process(
     if len(images) > 0:
         filtered_by_color = color_filtering_mask(
             time,
-            current_nodes,
-            next_nodes,
+            source_nodes,
+            target_nodes,
             images,
             neighbors,
             config.z_score_threshold,
@@ -182,23 +209,23 @@ def _process(
     else:
         filtered_by_color = np.ones_like(neighbors, dtype=bool)
 
-    int_next_shift = np.round(next_shift).astype(int)
+    int_next_shift = np.round(target_shift).astype(int)
     # NOTE: moving bbox with shift, MUST be after `feature computation`
-    for node, shift in zip(next_nodes, int_next_shift):
+    for node, shift in zip(target_nodes, int_next_shift):
         node.bbox[:n_dim] += shift
         node.bbox[-n_dim:] += shift
 
     distance_w = config.distance_weight
     links = []
 
-    for i, node in enumerate(next_nodes):
+    for i, node in enumerate(target_nodes):
         valid = (~np.isinf(distances[i])) & filtered_by_color[i]
         valid_neighbors = neighbors[i, valid]
         neigh_distances = distances[i, valid]
 
         neighborhood = []
         for neigh_idx, neigh_dist in zip(valid_neighbors, neigh_distances):
-            neigh = current_nodes[neigh_idx]
+            neigh = source_nodes[neigh_idx]
             edge_weight = node.IoU(neigh) - distance_w * neigh_dist
             # using dist as a tie-breaker
             neighborhood.append(
@@ -219,13 +246,14 @@ def _process(
 
     with write_lock if write_lock is not None else nullcontext():
         LOG.info(f"Pushing links from time {time} to {db_path}")
+        connect_args = {"timeout": 45} if write_lock is not None else {}
         engine = sqla.create_engine(
             db_path, hide_parameters=True, connect_args=connect_args
         )
         with engine.begin() as conn:
-            df.to_sql(
-                name=LinkDB.__tablename__, con=conn, if_exists="append", index=False
-            )
+            df.to_sql(name=table_name, con=conn, if_exists="append", index=False)
+
+    return df
 
 
 def link(
